el_vuelto_backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
python-decouple==3.8
Pillow==11.1.0
python-escpos==3.1
psycopg2-binary==2.9.10
